+++ .phan/plugins/DuplicateArrayKeyPlugin.php
@@ -1,10 +1,7 @@
 <?php declare(strict_types=1);
 
 use ast\Node;
-use Phan\AST\ContextNode;
-use Phan\Exception\CodeBaseException;
-use Phan\Exception\IssueException;
-use Phan\Exception\NodeException;
+use Phan\AST\UnionTypeVisitor;
 use Phan\Issue;
 use Phan\PluginV2;
 use Phan\PluginV2\PluginAwarePostAnalysisVisitor;
@@ -63,14 +60,22 @@ class DuplicateArrayKeyVisitor extends PluginAwarePostAnalysisVisitor
                 continue;  // This is `default:`. php --syntax-check already checks for duplicates.
             }
             // Skip array entries without literal keys. (Do it before resolving the key value)
-            $case_cond = $this->tryToResolveKey($case_cond);
-
             if (!is_scalar($case_cond)) {
-                // Skip non-literal keys.
-                continue;
+                $case_cond = UnionTypeVisitor::unionTypeFromNode($this->code_base, $this->context, $case_cond)->asSingleScalarValueOrNullOrSelf();
+                if (is_object($case_cond)) {
+                    // Skip non-literal keys.
+                    continue;
+                }
             }
-            if (isset($case_constant_set[$case_cond])) {
-                $normalized_case_cond = self::normalizeKey($case_cond);
+            if (is_string($case_cond)) {
+                $cond_key = "s$case_cond";
+            } elseif (is_int($case_cond)) {
+                $cond_key = $case_cond;
+            } else {
+                $cond_key = json_encode($case_cond);
+            }
+            if (isset($case_constant_set[$cond_key])) {
+                $normalized_case_cond = self::normalizeSwitchKey($case_cond);
                 $this->emitPluginIssue(
                     $this->code_base,
                     clone($this->context)->withLineNumberStart($case_node->lineno),
@@ -82,7 +87,7 @@ class DuplicateArrayKeyVisitor extends PluginAwarePostAnalysisVisitor
                     15071
                 );
             }
-            $case_constant_set[$case_cond] = true;
+            $case_constant_set[$cond_key] = $case_node->lineno;
         }
     }
 
@@ -114,12 +119,14 @@ class DuplicateArrayKeyVisitor extends PluginAwarePostAnalysisVisitor
                 $has_entry_without_key = true;
                 continue;
             }
-            $key = $this->tryToResolveKey($key);
-
             if (!is_scalar($key)) {
-                // Skip non-literal keys.
-                continue;
+                $key = UnionTypeVisitor::unionTypeFromNode($this->code_base, $this->context, $key)->asSingleScalarValueOrNullOrSelf();
+                if (is_object($key)) {
+                    // Skip non-literal keys.
+                    continue;
+                }
             }
+
             if (isset($key_set[$key])) {
                 $normalized_key = self::normalizeKey($key);
                 $this->emitPluginIssue(
@@ -133,6 +140,7 @@ class DuplicateArrayKeyVisitor extends PluginAwarePostAnalysisVisitor
                     15071
                 );
             }
+            // @phan-suppress-next-line PhanTypeMismatchDimAssignment
             $key_set[$key] = true;
         }
         if ($has_entry_without_key && count($key_set) > 0) {
@@ -150,43 +158,21 @@ class DuplicateArrayKeyVisitor extends PluginAwarePostAnalysisVisitor
     }
 
     /**
-     * @param int|string|float|array|Node $key (not actually an array)
-     * @return int|string|float|Node|array|bool|null - If possible, the value of $key converted to a scalar.
+     * Converts a key to the value it would be if used as a case.
+     * E.g. 0, 0.5, and "0" all become the same value(0) when used as an array key.
+     *
+     * @param int|string|float|bool|null $key - The array key literal to be normalized.
+     * @return string - The normalized representation.
      */
-    private function tryToResolveKey($key)
+    private static function normalizeSwitchKey($key) : string
     {
-        if (!($key instanceof ast\Node)) {
-            return $key;
+        if (is_int($key)) {
+            return (string)$key;
+        } elseif (!is_string($key)) {
+            return (string)json_encode($key);
         }
-        $kind = $key->kind;
-        if (!in_array($kind, [\ast\AST_CLASS_CONST, \ast\AST_CONST, \ast\AST_MAGIC_CONST], true)) {
-            return $key;
-        }
-        // if key is constant, take it in account
-        $context_node = new ContextNode($this->code_base, $this->context, $key);
-        try {
-            if ($kind === \ast\AST_CLASS_CONST) {
-                $key = $context_node->getClassConst()->getNodeForValue();
-            } elseif ($kind === \ast\AST_CONST) {
-                $key = $context_node->getConst()->getNodeForValue();
-            } else {
-                $key = $context_node->getValueForMagicConst();
-            }
-            return $key ?? '';
-        } catch (IssueException $e) {
-            // This is redundant, but do it anyway
-            Issue::maybeEmitInstance(
-                $this->code_base,
-                $this->context,
-                $e->getIssueInstance()
-            );
-        } catch (CodeBaseException $_) {
-            // e.g. Can't find the class (ignore)
-        } catch (NodeException $_) {
-            // E.g. Can't figure out constant class in node
-            // (ignore)
-        }
-        return $key;
+        $tmp = [$key => true];
+        return var_export(key($tmp), true);
     }
 
     /**
@@ -196,8 +182,11 @@ class DuplicateArrayKeyVisitor extends PluginAwarePostAnalysisVisitor
      * @param int|string|float|bool|null $key - The array key literal to be normalized.
      * @return string - The normalized representation.
      */
-    private static function normalizeKey($key) : string
+    private static function normalizeKey($key)
     {
+        if (is_int($key)) {
+            return (string)$key;
+        }
         $tmp = [$key => true];
         return var_export(key($tmp), true);
     }